enterprise_core/requirements.txt
fastapi
pydantic
uvicorn
redis
python-dotenv
PyYAML
aioredis
SQLAlchemy
psycopg2-binary
websockets
